Loading calibration files. Tested and working

+++ src/dataStructures.h
@@ -1,16 +1,11 @@
 #ifndef dataStructures_h
 #define dataStructures_h
 
-#include <vector>
 #include <opencv2/core.hpp>
+#include <vector>
 
 struct LidarPoint { // single lidar point in space
-    double x,y,z,r; // x,y,z in [m], r is point reflectivity
+  float x, y, z, r; // x,y,z in [m], r is point reflectivity
 };
 
-
 #endif /* dataStructures_h */
-
-
-
-
